Extract TVEpisode, TVObject, and TVShow to modules
